Hopefully fixed all 64 bit promotion issues - one last.
git-svn-id: http://svn.aircrack-ng.org/trunk@2441 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/tkiptun-ng.c
@@ -693,7 +693,7 @@ int read_packet(void *buf, size_t count, struct rx_info *ri)
     return rc;
 }
 
-void read_sleep( int usec )
+void read_sleep( unsigned long usec )
 {
     struct timeval tv, tv2, tv3;
     int caplen;
